% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccgSummaries.R
\name{ccgSummaries}
\alias{ccgSummaries}
\title{ccgSummaries}
\usage{
ccgSummaries(plpd, bnf, demog, demogMap, settings)
}
\arguments{
\item{plpd}{prescription data.table}

\item{bnf}{bnf data.table}

\item{demog}{demog data.table}

\item{demogMap}{demogMap data.table}

\item{settings}{analysis settings}
}
\description{
creates summaries at regional level
}
