% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDemogMap.R
\name{checkDemogMap}
\alias{checkDemogMap}
\title{checkDemogMap}
\usage{
checkDemogMap(demogMap, settings)
}
\arguments{
\item{demogMap}{settings with parameters}

\item{settings}{analysis settings}
}
\value{
The function returns the file corrected
}
\description{
Checks if demogMap is coherent for the specified month
}
\details{
For some months, the order and label of demogMap file has been found to be different.
This function checks if it is the case and correct it, if needed
}
