% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirsGen.R
\name{dirsGen}
\alias{dirsGen}
\title{dirsGen}
\usage{
dirsGen(settings)
}
\arguments{
\item{settings}{setting object created via \emph{setConfig()} function containing paths, month and year of analysis and file names.}
}
\value{
The function does not return anything but creates paths under the rootPath specified in the settings object.
 -
}
\description{
Reads paths and parameters from Settings and creates directories 
that will be used in the package for storing data and outputs
}
