% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData_Github.R
\name{downloadData_Github}
\alias{downloadData_Github}
\title{downloadData_Github}
\usage{
downloadData_Github(
  data = c("all", "bnf", "demog", "qof"),
  yyyy = 2019,
  mm = NULL,
  basedir = NULL
)
}
\arguments{
\item{data}{which data to download. One of "all","demog","bnf","qof", default = "all"}

\item{yyyy}{numerical 4 digit year, default = 2019}

\item{mm}{numerical 2 digit month, no default}

\item{basedir}{working directory, no default}
}
\value{
the function does not return anything, but download the specified files from github to the basedir folder
}
\description{
Download prescriptions data from PrescRiptionsData github repository where all data are saved
needed for the use of this R package
}
\details{
The function is used to download BNF (British National Formulary) codes, 
practice size for demographic groups and quality outcome framework data from the github repository
called PrescRiptionsData available at the following link:
https://github.com/muschitiello/PrescRiptionsData. 
While the repo contains data in different formats, only feather data are downloadable from this function.
This because feather data are more manageable and compressed.
}
\seealso{
\code{\link{importMonthlyData}}, \code{\link{downloadDemog}}, \code{\link{downloadPLPDzip}}
}
