% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPLPDzip.R
\name{downloadPLPDzip}
\alias{downloadPLPDzip}
\title{downloadPLPDzip}
\usage{
downloadPLPDzip(yyyy = 2019, mm = NULL, basedir = NULL, outFormat = "feather")
}
\arguments{
\item{yyyy}{numerical 4 digit year, default = 2019}

\item{mm}{numerical 2 digit month, no default}

\item{basedir}{working directory, no default}

\item{outFormat}{desired output format. One of "csv" or "feather", default = "feather"}
}
\description{
Download monthly plpd zip file for years 2018, 2019 in csv or feather format
and extract them on a specified folder.
}
