% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSummaries.R
\name{generateSummaries}
\alias{generateSummaries}
\title{generateSummaries}
\usage{
generateSummaries(plpd, bnf, on = "REGION", settings)
}
\arguments{
\item{plpd}{prescription data.table}

\item{bnf}{bnf data.table}

\item{on}{one of \emph{"REGION"} or \emph{"BNF"} indicating the variables to summarise on. Default is "REGION"}

\item{settings}{analysis settings}
}
\description{
creates summaries at regional level
}
