% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubDataDownload.R
\name{githubDataDownload}
\alias{githubDataDownload}
\title{githubDataDownload}
\usage{
githubDataDownload(
  data = c("all", "plpd", "demog", "bnf", "qof"),
  yyyy = 2019,
  mm = NULL,
  basedir = NULL,
  outFormat = "feather"
)
}
\arguments{
\item{data}{which data to download. One of "all","plpd","demog","bnf","qof", default = "all"}

\item{yyyy}{numerical 4 digit year, default = 2019}

\item{mm}{numerical 2 digit month, no default}

\item{basedir}{working directory, no default}

\item{outFormat}{desired output format. One of "csv", "feather" or "both", default = "feather"}
}
\description{
Download prescriptions data from PrescRiptionsData github repository where all data are saved
needed for the use of this R package
}
\details{
The function is used to download Practice Level PRescriprion data foor
years 2018 and 2019. Links for the download are specifyed internally.
The function checks if the file already exists and in that case, does NOT download it.
FIle are in csv format. Feather format can be chosen. In that case the csv file is converted and then removed, so to
save internal space. If "both" \emph{outFotmat} is selected, csv is kept and the feather is generated.
}
