% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpSummaries.R
\name{gpSummaries}
\alias{gpSummaries}
\title{gpSummaries}
\usage{
gpSummaries(plpd, demog, demogMap, settings)
}
\arguments{
\item{plpd}{practice level prescribing data.table}

\item{demog}{demog data.table}

\item{demogMap}{demogMap data.table}

\item{settings}{analysis settings generated with \emph{setConfig()}}
}
\value{
a summary data.table
}
\description{
creates aggregated values and calclate Cost rates at GP level
}
