% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMonthlyData.R
\name{importMonthlyData}
\alias{importMonthlyData}
\title{importMonthlyData}
\usage{
importMonthlyData(yyyy = 2019, mm = NULL, whichData = "all")
}
\arguments{
\item{yyyy}{numerical 4 digit year, default = 2019}

\item{mm}{numerical 2 digit month, no default}

\item{whichData}{string indicatin which Data to download. One of "plpd", "bnf", "demog" and "all"
default = "all"}
}
\value{
the function returns all data upoaded in the workspace
}
\description{
import in the workspace plpd, bnf, demog and qof data for the specified month and year as settend in the yml file
}
\details{
Import all data for one month analysis in the workspace. data are imported from the following sources:
 - plpd: website 
 - bnf: github/csv
 - demog: github/csv
 - qof: github/csv
github repository at this link: \emph{https://github.com/muschitiello/PrescRiptionsData}
}
\seealso{
\code{\link{downloadData_Github}}, \code{\link{downloadDemog}}, \code{\link{downloadPLPDzip}}
}
