% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyData_download.R
\name{monthlyData_download}
\alias{monthlyData_download}
\title{monthlyData_download}
\usage{
monthlyData_download(settings, whichData = "all")
}
\arguments{
\item{settings}{analysis settings}

\item{whichData}{string indicatin which Data to download. One of "plpd", "bnf", "demog" and "all"
default = "all"}
}
\value{
the function returns all data upoaded in the workspace
}
\description{
download plpd, bnf, demog and qof data in csv format for the specified month and year as settend in the yml file
}
\details{
Download all data for one month analysis. data are imported from the following sources:
 - plpd: website 
 - bnf: github/csv
 - demog: github/csv
 - qof: github/csv
github repository at this link: \emph{https://github.com/muschitiello/PrescRiptionsData}
}
\seealso{
\code{\link{monthlyData_import}}
}
