% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plpdDemogBridge.R
\name{plpdDemogBridge}
\alias{plpdDemogBridge}
\title{plpdDemogBridge}
\usage{
plpdDemogBridge(plpd = NULL, demogMap = NULL, settings = NULL)
}
\arguments{
\item{plpd}{prescription data.table}

\item{demogMap}{demogMap data.table}

\item{settings}{name of settings variable.}
}
\description{
creates summaries at regional level
}
