% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSettings.R
\name{readSettings}
\alias{readSettings}
\title{readSettings}
\usage{
readSettings(yamlF = NULL)
}
\arguments{
\item{yamlF}{character. Path of the settings file.}
}
\description{
A settings file allows a user of the package to make changes to the parameters of
a script without changing the script itself.
}
