% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderMonthlyReport.R
\name{renderMonthlyReport}
\alias{renderMonthlyReport}
\title{renderMonthlyReport}
\usage{
renderMonthlyReport(settings, monthData, ccg = NULL, gp = NULL, sample = F)
}
\arguments{
\item{settings}{settings with parameters}

\item{monthData}{list with all data for month, as returded by \emph{monthlyData_import()}}

\item{ccg}{CCG to run the analysis and the report for}

\item{gp}{GP to run the analysis and the report for}

\item{sample}{if sample data have been used, use this for making a message appearing in the report.}
}
\value{
The function generates a html file
}
\description{
Automatically generates a Markdown pdf report for a single month
}
\details{
If some encoding problem occurs, please set UTF-8 as saving enconding from \emph{File/Save with Encoding...}
}
