% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderReport.R
\name{renderReport}
\alias{renderReport}
\title{renderReport}
\usage{
renderReport(month, year = NULL, outFolder = settings$basedir)
}
\arguments{
\item{month}{month of interest.}

\item{year}{year of interest.}

\item{outFolder}{Folder where to save the Report}
}
\value{
The function generates a html file
}
\description{
Automatically generates a Markdown pdf report for a single month
}
\details{
If some encoding problem occurs, please set UTF-8 as saving enconding from \emph{File/Save with Encoding...}
}
