% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setConfig.R
\name{setConfig}
\alias{setConfig}
\title{setConfig}
\usage{
setConfig(basedir, year, month, region)
}
\arguments{
\item{basedir}{path to an existing working directory}

\item{year}{numeric. Year of interest}

\item{month}{numeric. Month of interest}

\item{region}{char. Region of analysis}
}
\description{
creates a list of configuration settings that will be used across the package
}
