% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureParameters.R
\name{structureParameters}
\alias{structureParameters}
\title{structureParameters}
\usage{
structureParameters(settings = NULL)
}
\arguments{
\item{settings}{analysis settings as uploaded form a pesi.yml file}
}
\value{
folder & parameters as a list
}
\description{
structureParameters
}
