% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_params.R
\name{w_params}
\alias{w_params}
\title{w_params}
\usage{
w_params(ymlp = NULL)
}
\arguments{
\item{ymlp}{path tp yml parameters file}
}
\description{
Generate working parameters from an external YML file set by the user
}
