% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSettings.R
\name{addSettings}
\alias{addSettings}
\title{addSettings}
\usage{
addSettings(dir = ".")
}
\arguments{
\item{dir}{character. Directory in which to write settings files}
}
\description{
Creates a sample settings file, by default (leading dashes for
  formatting, whitespace in the actual file):

year: [year] \cr
month: [month] \cr
basedir: [working directory] \cr
}
\details{
Add settings file to project
}
