% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPLPDzip.R
\name{downloadPLPDzip}
\alias{downloadPLPDzip}
\title{downloadPLPDzip}
\usage{
downloadPLPDzip(yyyy = 2019, mm = 12, basedir = "D:/Progetti/2020_erum2020/")
}
\arguments{
\item{yyyy}{numerical 4 digit year}

\item{mm}{numerical 2 digit month}

\item{basedir}{working directory}
}
\description{
Download monthly plpd zip file for years 2018, 2019 and extract them on a specified folder
}
