% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPLPDzip.R
\name{downloadPLPDzip}
\alias{downloadPLPDzip}
\title{downloadPLPDzip}
\usage{
downloadPLPDzip(yyyy = 2019, mm = NULL, basedir = NULL, outFormat = "feather")
}
\arguments{
\item{yyyy}{numerical 4 digit year, default = 2019}

\item{mm}{numerical 2 digit month, no default}

\item{basedir}{working directory, no default}

\item{outFormat}{desired output format. One of "csv", "feather" or "both", default = "feather"}
}
\description{
Download monthly plpd zip file for years 2018, 2019 in csv or feather format
and extract them on a specified folder.
}
\details{
The function is used to download Practice Level PRescriprion data foor
years 2018 and 2019. Links for the download are specifyed internally.
The function checks if the file already exists and in that case, does NOT download it.
FIle are in csv format. Feather format can be chosen. In that case the csv file is converted and then removed, so to
save internal space. If "both" \emph{outFotmat} is selected, csv is kept and the feather is generated.
}
