% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSummaries.R
\name{generateSummaries}
\alias{generateSummaries}
\title{generateSummaries}
\usage{
generateSummaries(
  plpd = NULL,
  bnf = NULL,
  on = "REGION",
  year = NULL,
  month = NULL
)
}
\arguments{
\item{plpd}{prescription data.table}

\item{bnf}{bnf data.table}

\item{on}{one of \emph{"REGION"} or \emph{"BNF"} indicating the variables to summarise on. Default is "REGION"}

\item{year}{year of interest.}

\item{month}{month of interest.}
}
\description{
creates summaries at regional level
}
