% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyData_download.R
\name{monthlyData_download}
\alias{monthlyData_download}
\title{monthlyData_download}
\usage{
monthlyData_download(settings, whichData = "all", sample)
}
\arguments{
\item{settings}{setting object created via \emph{setConfig()} function containing paths, month and year of analysis and file names.}

\item{whichData}{string indicatin which Data to download. One of "plpd", "bnf", "demog" and "all"
default = "all"}

\item{sample}{logic. if TRUE (default) a sample file is downloaded from github Repository with only 500.000 rows.
When FALSE completa data are downloaded from website.}
}
\value{
the function does not return anything but downloads data in the \emph{dataInput} folder and under the relative subfolder
}
\description{
Download plpd, bnf, demog and qof data in csv format and save them in the data directory as generated by the \emph{dirsGen()} function
}
\details{
Verify if data have already been downloaded and, if not, download all data for one month analysis. 
Data are imported from the following sources:
 - plpd: website 
 - bnf: github/csv
 - demog: github/csv
 - qof: github/csv
github repository at this link: \emph{https://github.com/muschitiello/PrescRiptionsData}
For time saving purposes the function checks if data are already available in the relative path.
If the data have already been downloaded, a message is returned saying that data are already available.
}
\seealso{
\code{\link{monthlyData_import}}
}
