% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyData_import.R
\name{monthlyData_import}
\alias{monthlyData_import}
\title{monthlyData_import}
\usage{
monthlyData_import(settings, whichData = "all")
}
\arguments{
\item{settings}{analysis settings}

\item{whichData}{string indicatin which Data to download. One of "plpd", "bnf", "demog" and "all"
default = "all"}
}
\value{
the function returns all data upoaded in the workspace
}
\description{
import in the workspace plpd, bnf, demog and qof data for the specified month and year as settend in the yml file
}
\details{
Import all data for one month analysis in the workspace. data are imported from the following sources:
 - plpd: website 
 - bnf: github/csv
 - demog: github/csv
 - qof: github/csv
github repository at this link: \emph{https://github.com/muschitiello/PrescRiptionsData}
}
\seealso{
\code{\link{monthlyData_download}}
}
