% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyData_import.R
\name{monthlyData_import}
\alias{monthlyData_import}
\title{monthlyData_import}
\usage{
monthlyData_import(settings, whichData = "all", sample)
}
\arguments{
\item{settings}{analysis settings generated with \emph{setConfig()}}

\item{whichData}{string indicatin which Data to download. One of "plpd", "bnf", "demog" and "all"
default = "all"}

\item{sample}{logic. if TRUE (default) only 500.000 rows will be uploaded in the workspace.}
}
\value{
the function returns all data upoaded in the workspace in list format
}
\description{
import plpd, bnf, demog and qof data in the workspace for the specified month and year as setted with \emph{setConfig()}
}
\details{
Import all data for one month analysis in the workspace. If altready downloaded, data are retrived from the folder where they have been downloaded.
If not already downloaded, data are downloaded from:
 - plpd: website if sample = FALSE, github if sample = TRUE (default)
 - bnf: github/csv
 - demog: github/csv
 - qof: github/csv
github repository at this link: \emph{https://github.com/muschitiello/PrescRiptionsData}
}
\seealso{
\code{\link{monthlyData_download}}
}
