% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plpdDemogBridge.R
\name{plpdDemogBridge}
\alias{plpdDemogBridge}
\title{plpdDemogBridge}
\usage{
plpdDemogBridge(plpd = NULL, demogMap = NULL, settings = NULL)
}
\arguments{
\item{plpd}{prescription data.table}

\item{demogMap}{demogMap data.table}

\item{settings}{analysis settings generated with \emph{setConfig()}}
}
\value{
the function returns a data table composed on 2 columns, one for PRACTICE code and one for CCG code. 
 Moreover, a message is returned, with the number of unmatched records.
}
\description{
links PRACTICE codes and CCG codes
}
\details{
Some records are not matched and this is due to the fact that PRactice level Prescribing data are more "generic" and include also GP that are not mapped by the CCG system.
}
