% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderMonthlyReport.R
\name{renderMonthlyReport}
\alias{renderMonthlyReport}
\title{renderMonthlyReport}
\usage{
renderMonthlyReport(settings, outFolder)
}
\arguments{
\item{settings}{settings with parameters}

\item{outFolder}{Folder where to save the Report}
}
\value{
The function generates a html file
}
\description{
Automatically generates a Markdown pdf report for a single month
}
\details{
If some encoding problem occurs, please set UTF-8 as saving enconding from \emph{File/Save with Encoding...}
}
