% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderMonthlyReport.R
\name{renderMonthlyReport}
\alias{renderMonthlyReport}
\title{renderMonthlyReport}
\usage{
renderMonthlyReport(settings, monthData, ccg = NULL, gp = NULL)
}
\arguments{
\item{settings}{analysis settings generated with \emph{setConfig()}}

\item{monthData}{list with all data for month, as returded by \emph{monthlyData_import()}}

\item{ccg}{In case a focus is neded for a specific CCG, the CCG to run the analysis and the report for.}

\item{gp}{In case a focus is neded for a specific GP, the GP to run the analysis and the report for}
}
\value{
The function generates an html file.
}
\description{
Automatically generates a Markdown pdf report for a single month.
}
\details{
The fucntion generate a monthly report composed of different sections: 
  - Report by Region: summary aggregations at England Level for the selected month.
  - Analysis by BNF codes: aggregations at BNF Code Level for the selected month.
  - Cost Summary by CCG: Net Ingredient cost for 1K patiensts in England for the selected month.
  - Costlier GPs:  list of practices who were in the higher end of the distribution of Net Ingredient Cost per 1K patients 
  - Report by CCG (customised section): aggregation for a single CCG, if selected.
  - Report by GP (customised section): aggregations for a single GP, if selected.
  If some encoding problem occurs, please set UTF-8 as saving enconding from \emph{File/Save with Encoding...}
}
