% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setConfig.R
\name{setConfig}
\alias{setConfig}
\title{setConfig}
\usage{
setConfig(year, month, basedir)
}
\arguments{
\item{year}{numeric. Year of interest}

\item{month}{numeric. Month of interest}

\item{basedir}{path to an existing working directory}
}
\description{
creates a list of configuration settings that will be used across the package
}
