% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setConfig.R
\name{setConfig}
\alias{setConfig}
\title{setConfig}
\usage{
setConfig(rootdir, year, month, region)
}
\arguments{
\item{rootdir}{path to an existing working directory}

\item{year}{numeric. Year of interest}

\item{month}{numeric. Month of interest, one of 2018 and 2019}

\item{region}{char. Region of analysis}
}
\value{
The function returns a list containing the following information, as relative to the month and year selected:
 - rootdir
 - year
 - month
 - region
 - plpdName
 - demogName
 - demogMapName 
 - bnfName
 - qofNames
}
\description{
creates a list of configuration settings that will be used across the package
}
